pull cache into separate file
