+++ src/packs/caching/per_file_cache.rs
@@ -6,7 +6,7 @@ use std::io::Write;
 use std::path::Path;
 use std::path::PathBuf;
 
-use super::Cache;
+use super::cache::Cache;
 use super::CacheResult;
 use super::EmptyCacheEntry;
